requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
gradio>=4.0.0
scikit-learn
matplotlib
Pillow
numpy
pandas
tqdm
albumentations
opencv-python
pydantic-settings
python-dotenv
